+++ src/num/prob/mod.rs
@@ -1,7 +1,7 @@
 // devela/src/num/prob/mod.rs
 //
 #![doc = crate::_DOC_NUM_PROB!()] // public
-#![doc = crate::_doc!(modules: crate::num; prob: dist, rand)]
+#![doc = crate::_doc!(modules: crate::num; prob: dist, rand, stats)]
 #![doc = crate::_doc!(flat:"num")]
 #![doc = crate::_doc!(hr)]
 //!
@@ -14,7 +14,7 @@ pub mod dist; // Probability distributions
 // mod markov; // Stochastic matrices, chains, traces
 mod probability; // Probability
 pub mod rand; // Random number generation
-// pub mod stats; // Descriptive statistics
+pub mod stats; // Descriptive statistics
 
 crate::structural_mods! { // _mods, _pub_mods, _crate_internals, _hidden
     _mods {
@@ -29,7 +29,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals, _hidden
             dist::_all::*,
             // markov::_all::*,
             rand::_all::*,
-            // stats::_all::*,
+            stats::_all::*,
         };
     }
     _crate_internals {